refactored component with information like title, name and date
