+++ src/layouts/Tile/movies/movieDetails/index.js
@@ -1,6 +1,6 @@
 import Description from "../../Description";
 import Details from "../../Details";
-import { Information } from "../../Information";
+import Information from "../../Information";
 import { Raiting } from "../../Raiting";
 import { Tags } from "../../Tags";
 import { Container, DescriptionContainer, Image, MovieContainer } from "./styled";
