+++ src/layouts/Tile/persons/personDetails/index.js
@@ -1,6 +1,6 @@
 import Description from "../../Description";
 import Details from "../../Details";
-import { Information } from "../../Information";
+import Information from "../../Information";
 import { Container, DescriptionContainer, Image } from "./styled";
 import noPerson from "../../../../common/images/no-person.png";
 import { APIImageUrl } from "../../../../features/getAPI";
